# openapi.yaml
openapi: 3.1.0
info:
  title: Community Platform API
  description: |
    Complete REST API for a Next.js community platform with social features, vendor marketplace, 
    bookings, gamification, and admin capabilities.
    
    ## Features
    - 🔐 Authentication via Clerk
    - 🛡️ Rate limiting with Upstash Redis
    - ✅ Zod validation on all inputs
    - 📊 Supabase PostgreSQL with RLS
    - 💳 Stripe payment integration
    - 📧 Webhook support
    
    ## Authentication
    All authenticated endpoints require a valid Clerk session token in the Authorization header:
    ```
    Authorization: Bearer <clerk_jwt_token>
    ```
    
    ## Rate Limiting
    Rate limits vary by endpoint and authentication status:
    - Anonymous read: 20 requests/minute
    - Anonymous write: 5 requests/minute
    - Authenticated read: 100 requests/minute
    - Authenticated write: 30 requests/minute
    - Admin operations: 10 requests/minute
    - Webhooks: 100 requests/minute
    
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Timestamp when limit resets
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": "Error message",
      "details": "Additional error details or validation errors"
    }
    ```
    
    Common HTTP status codes:
    - `400` Bad Request - Invalid input or validation error
...
    
...
servers:
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clerk JWT token obtained from user authentication.
        Include in Authorization header: `Authorization: Bearer <token>`
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Validation failed
        details:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
          description: Additional error details or validation errors
          example:
            - field: content
              message: String must contain at least 1 character(s)
...
    RateLimitError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Rate limit exceeded
        retryAfter:
          type: integer
          description: Seconds until rate limit resets
          example: 45
...
    ProfileSummary:
      type: object
      required:
        - id
        - username
      properties:
        id:
          type: string
          format: uuid
          description: Profile ID
        username:
          type: string
          description: User's username
        display_name:
          type: string
          description: User's display name
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Profile picture URL
        is_verified:
          type: boolean
          description: User verification status
        vendor_verified:
          type: boolean
          description: Vendor verification status
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized. Please sign in.
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests allowed
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      description: User UUID or Clerk ID
      schema:
        type: string
      example: 123e4567-e89b-12d3-a456-426614174000
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of items to skip (for pagination)
      schema:
        type: integer
        minimum: 0
        default: 0
paths:
...
  /api/users/{id}/follow:
    get:
      tags:
        - Follow
      summary: Get followers or following list
      description: |
        Get list of users who follow this user (followers) or users this user follows (following).
        Use query parameter 'type' to specify which list to retrieve.
      operationId: getFollowList
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: type
          in: query
          description: Type of list to retrieve
          schema:
            type: string
            enum: [followers, following]
            default: followers
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Follow list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileSummary'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  type:
                    type: string
                    enum: [followers, following]
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Follow
      summary: Follow a user
      description: |
        Follow a user. Requires authentication. Creates notification for target user.
        Prevents self-follows. Updates follower/following counts.
      operationId: followUser
      security:
        - ClerkAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User followed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User followed successfully
        '400':
          description: Already following or attempted self-follow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                alreadyFollowing:
                  value:
                    error: Already following this user
                selfFollow:
                  value:
                    error: Cannot follow yourself
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Follow
      summary: Unfollow a user
      description: |
        Unfollow a user. Requires authentication.
        Updates follower/following counts.
      operationId: unfollowUser
      security:
        - ClerkAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User unfollowed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User unfollowed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
security:
  - ClerkAuth: []